web/requirements.txt
streamlit
matplotlib
pyyaml
numpy
